[UIManager] fix #745
ui.remove now squals an IllegalArgumentException if the file is not there
to remove.

git-svn-id: https://bioclipse.svn.sourceforge.net/svnroot/bioclipse/bioclipse2/trunk@10343 fcb5ba71-d80d-0410-8237-ba3920747fcc

+++ plugins/net.bioclipse.ui.business/src/net/bioclipse/ui/business/UIManager.java
@@ -76,6 +76,10 @@ public class UIManager implements IUIManager {
 
     public void remove( IFile file, IProgressMonitor monitor ) {
         try {
+            if (!file.exists())
+                throw new IllegalArgumentException(
+                    "File not found: " + file.getName()
+                );
             file.delete(true, monitor);
         } catch (PartInitException e) {
             throw new RuntimeException(e);
